+++ src/components/splitter/types.ts
@@ -37,18 +37,32 @@ interface IgcSplitterResizeEventArgs {
 
 /**
  * @hidden
- * @deprecated use IgcSplitterResizeEventArgs instead
+ * @deprecated since 7.1.0. Use the `IgcSplitterResizeEventArgs` type instead.
  */
 interface IgcSplitterResizeEventDetail extends IgcSplitterResizeEventArgs {}
+
+/* jsonAPIPlainObject */
+interface IgcSplitterLayoutChangedEventArgs {
+  /** The current size of the start pane */
+  startSize: string;
+  /** The current size of the end pane */
+  endSize: string;
+  /** Whether the start pane is currently collapsed */
+  startCollapsed: boolean;
+  /** Whether the end pane is currently collapsed */
+  endCollapsed: boolean;
+}
 
 interface IgcSplitterComponentEventMap {
   igcResizeStart: CustomEvent<IgcSplitterResizeEventArgs>;
   igcResizing: CustomEvent<IgcSplitterResizeEventArgs>;
   igcResizeEnd: CustomEvent<IgcSplitterResizeEventArgs>;
+  igcLayoutChanged: CustomEvent<IgcSplitterLayoutChangedEventArgs>;
 }
 
 export type {
   IgcSplitterComponentEventMap,
+  IgcSplitterLayoutChangedEventArgs,
   IgcSplitterResizeEventArgs,
   IgcSplitterResizeEventDetail,
   PanePosition,